% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{accoustic}
\alias{accoustic}
\title{Tukish Music Emotion Data Set}
\format{
A data frame with 400 observations and 51 features
}
\source{
Bilal Er, M., & Aydilek, I. B. (2019).
   Music emotion recognition by using chroma spectrogram and deep visual
   features.
   Journal of Computational Intelligent Systems, 12(2), 1622–1634.
   International Journal of Computational Intelligence Systems,
   DOI: https://doi.org/10.2991/ijcis.d.191216.001
}
\usage{
data(accoustic)
}
\description{
a database of music for happy, sad, angry, relax emotions prepared by
selection of verbal and non-verbal music  from different genres of
Turkish music. The dataset is designed as a discrete model,
and there are four classes. A total of 100 music pieces are determined for
each class in the database to have an equal number of samples in each class.
There are 400 samples in the original dataset as 30 seconds from each sample.
Happy 100, Sad 100, Angry 100, Relax 100
}
