% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ackley.R
\name{ackley}
\alias{ackley}
\title{ACKLEY FUNCTION}
\usage{
ackley(xx, a = 20, b = 0.2, c = 2 * pi)
}
\arguments{
\item{xx}{c(x1, x2, ..., xd)}

\item{a}{constant (optional), with default value 20}

\item{b}{constant (optional), with default value 0.2}

\item{c}{constant (optional), with default value 2*pi}
}
\description{
Dimensions: d
	The Ackley function is widely used for testing optimization algorithms. In its two-dimensional form, as shown in the plot above, it is characterized by a nearly flat outer region, and a large hole at the centre. The function poses a risk for optimization algorithms, particularly hillclimbing algorithms, to be trapped in one of its many local minima.
	Recommended variable values are: a = 20, b = 0.2 and c = 2p.
	The function is usually evaluated on the hypercube xi <U+2208> [-32.768, 32.768], for all i = 1, ?, d, although it may also be restricted to a smaller domain.
}
\references{
\url{https://www.sfu.ca/~ssurjano/ackley.html}
	Adorio, E. P., & Diliman, U. P. MVF - Multivariate Test Functions Library in C for Unconstrained Global Optimization (2005). Retrieved June 2013, from http://http://www.geocities.ws/eadorio/mvf.pdf.
	Molga, M., & Smutnicki, C. Test functions for optimization needs (2005). Retrieved June 2013, from http://www.zsd.ict.pwr.wroc.pl/files/docs/functions.pdf.
	Back, T. (1996). Evolutionary algorithms in theory and practice: evolution strategies, evolutionary programming, genetic algorithms. Oxford University Press on Demand.
}
\author{
Sonja Surjanovic, Simon Fraser University
         	Derek Bingham, Simon Fraser University
}
