% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_budget}
\alias{add_budget}
\title{budget addition}
\usage{
add_budget(object, its, new_budget)
}
\arguments{
\item{object}{an egoOptim object or a list of egoOptim objects}

\item{its}{the new number of iterations}

\item{new_budget}{the number of points to be added}
}
\description{
Uses EGO algorithm to optimize any given function.
}
\author{
BLANK
}
