% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_replicates}
\alias{add_replicates}
\title{Replications addition}
\usage{
add_replicates(object, reps)
}
\arguments{
\item{object}{a list of egoObjects}

\item{reps}{the number of replicates to be added}
}
\description{
Replications addition
}
\author{
None
}
