% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beale.R
\name{beale}
\alias{beale}
\title{BEALE FUNCTION}
\usage{
beale(xx)
}
\arguments{
\item{xx}{c(x1, x2)}
}
\description{
Dimensions: 2
	The Beale function is multimodal, with sharp peaks at the corners of the input domain.
	The function is usually evaluated on the square xi <U+2208> [-4.5, 4.5], for all i = 1, 2.
	Reference:
	http://www-optima.amp.i.kyoto-u.ac.jp/member/student/hedar/Hedar_files/TestGO.htm.
}
\author{
Sonja Surjanovic, Simon Fraser University
         	Derek Bingham, Simon Fraser University
}
