% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boha3.R
\name{boha3}
\alias{boha3}
\title{BOHACHEVSKY FUNCTION 3}
\usage{
boha3(xx)
}
\arguments{
\item{xx}{c(x1, x2)}
}
\description{
Dimensions: 2
	The Bohachevsky functions all have the same similar bowl shape. The one shown above is the first function.
	The functions are usually evaluated on the square xi <U+2208> [-100, 100], for all i = 1, 2.
	Reference:
	http://www-optima.amp.i.kyoto-u.ac.jp/member/student/hedar/Hedar_files/TestGO.htm.
}
\author{
Sonja Surjanovic, Simon Fraser University
         	Derek Bingham, Simon Fraser University
}
