% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{breastTissue}
\alias{breastTissue}
\alias{breast_tissue}
\title{Breast Tissue}
\format{
A data frame with 106 observations and 10 features
}
\source{
S,JP and Jossinet,J. (2010). Breast Tissue.
 UCI Machine Learning Repository. https://doi.org/10.24432/C5P31H.
}
\usage{
data(breastTissue)
}
\description{
Impedance measurements were made at the frequencies: 15.625,
31.25, 62.5, 125, 250, 500, 1000 KHz. Impedance measurements of freshly
excised breast tissue were made at the follwoing frequencies:
 15.625, 31.25, 62.5, 125, 250, 500, 1000 KHz. These measurements plotted
  in the (real, -imaginary) plane constitute the impedance spectrum from
  where the breast tissue features are computed. The dataset can be used
  for predicting the classification of either the original 6 classes or of
  4 classes by merging together the fibro-adenoma, mastopathy and glandular
  classes whose discrimination is not important (they cannot be accurately
   discriminated anyway).
}
\details{
\describe{
\item{\code{Class}}{The response variable with 6 levels:
  car(carcinoma), fad (fibro-adenoma), mas (mastopathy), gla (glandular),
  con (connective), adi (adipose)}
\item{\code{I0}}{Impedivity (ohm) at zero frequency}
\item{\code{PA500}}{phase angle at 500 KHz}
\item{\code{HFS}}{high-frequency slope of phase angle}
\item{\code{DA}}{impedance distance between spectral ends}
\item{\code{AREA}}{area under spectrum}
\item{\code{A/DA}}{area normalized by DA}
\item{\code{MAX IP}}{maximum of the spectrum}
\item{\code{DR}}{distance between I0 and real part of the
     maximum frequency point}
\item{\code{P}}{length of the spectral curve}
}
}
