% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bukin6.R
\name{bukin6}
\alias{bukin6}
\title{BUKIN FUNCTION N. 6}
\usage{
bukin6(xx)
}
\arguments{
\item{xx}{c(x1, x2)}
}
\description{
Dimensions: 2
	The sixth Bukin function has many local minima, all of which lie in a ridge.
	The function is usually evaluated on the rectangle x1 <U+2208> [-15, -5], x2 <U+2208> [-3, 3].
	Reference:
}
\author{
Sonja Surjanovic, Simon Fraser University
         	Derek Bingham, Simon Fraser University
}
