% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camel3.R
\name{camel3}
\alias{camel3}
\title{THREE-HUMP CAMEL FUNCTION}
\usage{
camel3(xx)
}
\arguments{
\item{xx}{c(x1, x2)}
}
\description{
Dimensions: 2
	The plot on the left shows the three-hump Camel function on its recommended input domain, and the plot on the right shows only a portion of this domain, to allow for easier viewing of the function's key characteristics. The function has three local minima.
	The function is usually evaluated on the square xi <U+2208> [-5, 5], for all i = 1, 2.
	Reference:
	Test functions for optimization. In Wikipedia. Retrieved June 2013, from https://en.wikipedia.org/wiki/Test_functions_for_optimization.
}
\author{
Sonja Surjanovic, Simon Fraser University
         	Derek Bingham, Simon Fraser University
}
