% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camel6.R
\name{camel6}
\alias{camel6}
\title{SIX-HUMP CAMEL FUNCTION}
\usage{
camel6(xx)
}
\arguments{
\item{xx}{c(x1, x2)}
}
\description{
Dimensions: 2
	The plot on the left shows the six-hump Camel function on its recommended input domain, and the plot on the right shows only a portion of this domain, to allow for easier viewing of the function's key characteristics. The function has six local minima, two of which are global.
	The function is usually evaluated on the rectangle x1 <U+2208> [-3, 3], x2 <U+2208> [-2, 2].
	Reference:
	Molga, M., & Smutnicki, C. Test functions for optimization needs (2005). Retrieved June 2013, from http://www.zsd.ict.pwr.wroc.pl/files/docs/functions.pdf.
}
\author{
Sonja Surjanovic, Simon Fraser University
         	Derek Bingham, Simon Fraser University
}
