% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colville.R
\name{colville}
\alias{colville}
\title{COLVILLE FUNCTION}
\usage{
colville(xx)
}
\arguments{
\item{xx}{c(x1, x2, x3, x4)}
}
\description{
Dimensions: 4
	The function is usually evaluated on the hypercube xi <U+2208> [-10, 10], for all i = 1, 2, 3, 4.
	Reference:
	http://www-optima.amp.i.kyoto-u.ac.jp/member/student/hedar/Hedar_files/TestGO.htm.
}
\author{
Sonja Surjanovic, Simon Fraser University
         	Derek Bingham, Simon Fraser University
}
