% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossit.R
\name{crossit}
\alias{crossit}
\title{CROSS-IN-TRAY FUNCTION}
\usage{
crossit(xx)
}
\arguments{
\item{xx}{c(x1, x2)}
}
\description{
Dimensions: 2
	The Cross-in-Tray function has multiple global minima. It is shown here with a smaller domain in the second plot, so that its characteristic "cross" will be visible.
	The function is usually evaluated on the square xi <U+2208> [-10, 10], for all i = 1, 2.
	Reference:
	Test functions for optimization. In Wikipedia. Retrieved June 2013, from https://en.wikipedia.org/wiki/Test_functions_for_optimization.
}
\author{
Sonja Surjanovic, Simon Fraser University
         	Derek Bingham, Simon Fraser University
}
