% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dejong5.R
\name{dejong5}
\alias{dejong5}
\title{DE JONG FUNCTION N. 5}
\usage{
dejong5(xx)
}
\arguments{
\item{xx}{c(x1, x2)}
}
\description{
Dimensions: 2
	The fifth function of De Jong is multimodal, with very sharp drops on a mainly flat surface.
	The function is usually evaluated on the square xi <U+2208> [-65.536, 65.536], for all i = 1, 2.
	Reference:
	Molga, M., & Smutnicki, C. Test functions for optimization needs (2005). Retrieved June 2013, from http://www.zsd.ict.pwr.wroc.pl/files/docs/functions.pdf.
}
\author{
Sonja Surjanovic, Simon Fraser University
         	Derek Bingham, Simon Fraser University
}
