% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dixonpr.R
\name{dixonpr}
\alias{dixonpr}
\title{DIXON-PRICE FUNCTION}
\usage{
dixonpr(xx)
}
\arguments{
\item{xx}{c(x1, x2, ..., xd)}
}
\description{
Dimensions: d
	The function is usually evaluated on the hypercube xi <U+2208> [-10, 10], for all i = 1, ?, d.
	Reference:
}
\author{
Sonja Surjanovic, Simon Fraser University
         	Derek Bingham, Simon Fraser University
}
