% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.R
\name{drop}
\alias{drop}
\title{DROP-WAVE FUNCTION}
\usage{
drop(xx)
}
\arguments{
\item{xx}{c(x1, x2)}
}
\description{
Dimensions: 2
	The Drop-Wave function is multimodal and highly complex. The second plot above shows the function on a smaller input domain, to illustrate its characteristic features.
	The function is usually evaluated on the square xi <U+2208> [-5.12, 5.12], for all i = 1, 2.
	Reference:
}
\author{
Sonja Surjanovic, Simon Fraser University
         	Derek Bingham, Simon Fraser University
}
