% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easom.R
\name{easom}
\alias{easom}
\title{EASOM FUNCTION}
\usage{
easom(xx)
}
\arguments{
\item{xx}{c(x1, x2)}
}
\description{
Dimensions: 2
	The Easom function has several local minima. It is unimodal, and the global minimum has a small area relative to the search space.
	The function is usually evaluated on the square xi <U+2208> [-100, 100], for all i = 1, 2.
	Reference:
	http://www-optima.amp.i.kyoto-u.ac.jp/member/student/hedar/Hedar_files/TestGO.htm.
}
\author{
Sonja Surjanovic, Simon Fraser University
         	Derek Bingham, Simon Fraser University
}
