% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg.R
\name{egg}
\alias{egg}
\title{EGGHOLDER FUNCTION}
\usage{
egg(xx)
}
\arguments{
\item{xx}{c(x1, x2)}
}
\description{
Dimensions: 2
	The Eggholder function is a difficult function to optimize, because of the large number of local minima.
	The function is usually evaluated on the square xi <U+2208> [-512, 512], for all i = 1, 2.
	Reference:
}
\author{
Sonja Surjanovic, Simon Fraser University
         	Derek Bingham, Simon Fraser University
}
