% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{egoApply}
\alias{egoApply}
\title{Used to compute the means and medians of the resulting egoOptim replications}
\usage{
egoApply(object, ...)
}
\arguments{
\item{object}{list of list of egoObjects}

\item{...}{extra arguments such as the FUN to be evaluated.}
}
\description{
Used to compute the means and medians of the resulting egoOptim replications
}
