% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ESLmixture}
\alias{ESLmixture}
\title{ESL mixture data:}
\usage{
data(ESLmixture)
}
\description{
Data from essentiasls of statistical learning
}
\details{
has the points in \code{X} matrix with the labels in \code{y} vector
}
