% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{glass}
\alias{glass}
\title{Glass Identification}
\format{
A data frame with 214 observations and 10 features
}
\source{
German,B.. (1987). Glass Identification.
UCI Machine Learning Repository. https://doi.org/10.24432/C5WW2P.
}
\usage{
data(glass)
}
\description{
Impedance measurements were made at the frequencies: 15.625,
31.25, 62.5, 125, 250, 500, 1000 KHz. Impedance measurements of freshly
excised breast tissue were made at the follwoing frequencies:
 15.625, 31.25, 62.5, 125, 250, 500, 1000 KHz. These measurements plotted
  in the (real, -imaginary) plane constitute the impedance spectrum from
  where the breast tissue features are computed. The dataset can be used
  for predicting the classification of either the original 6 classes or of
  4 classes by merging together the fibro-adenoma, mastopathy and glandular
  classes whose discrimination is not important (they cannot be accurately
   discriminated anyway).
}
\details{
\describe{
\item{\code{type}}{class attribute with 7 levels \itemize{
	\item{1}{building_windows_float_processed}
\item{2}{building_windows_non_float_processed}
\item{3}{vehicle_windows_float_processed}
\item{4}{vehicle_windows_non_float_processed (none in this database)}
\item{5}{containers}
\item{6}{tableware}
\item{7}{headlamps}
}}
\item{\code{RI}}{refractive index}
\item{\code{measurement}}{weight percent in corresponding oxide, as are
attributes 4-10)}
\item{\code{Mg}}{Magnesium}
\item{\code{Al}}{Aluminum}
\item{\code{Si}}{Silicon}
\item{\code{K}}{Potassium}
\item{\code{Ca}}{Calcium}
\item{\code{Ba}}{Barium}
\item{\code{Fe}}{Iron}
}
}
