% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goldpr.R
\name{goldpr}
\alias{goldpr}
\title{GOLDSTEIN-PRICE FUNCTION}
\usage{
goldpr(xx)
}
\arguments{
\item{xx}{c(x1, x2)}
}
\description{
Dimensions: 2
	The Goldstein-Price function has several local minima.
	The function is usually evaluated on the square xi <U+2208> [-2, 2], for all i = 1, 2.
	Modifications and Alternate Forms:
	Picheny et al. (2012) use the following logarithmic form of the Goldstein-Price function, on [0, 1]2:
	This rescaled form of the function has a mean of zero and a variance of one. The authors also add a small Gaussian error term to the output.
}
\references{
\url{https://www.sfu.ca/~ssurjano/goldpr.html}
	Dixon, L. C. W., & Szego, G. P. (1978). The global optimization problem: an introduction. Towards global optimization, 2, 1-15.
	Molga, M., & Smutnicki, C. Test functions for optimization needs (2005). Retrieved June 2013, from http://www.zsd.ict.pwr.wroc.pl/files/docs/functions.pdf.
	Picheny, V., Wagner, T., & Ginsbourger, D. (2012). A benchmark of kriging-based infill criteria for noisy optimization.
}
\author{
Sonja Surjanovic, Simon Fraser University
         	Derek Bingham, Simon Fraser University
}
