% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/griewank.R
\name{griewank}
\alias{griewank}
\title{GRIEWANK FUNCTION}
\usage{
griewank(xx)
}
\arguments{
\item{xx}{c(x1, x2, ..., xd)}
}
\description{
Dimensions: d
	The Griewank function has many widespread local minima, which are regularly distributed. The complexity is shown in the zoomed-in plots.
	The function is usually evaluated on the hypercube xi <U+2208> [-600, 600], for all i = 1, ?, d.
}
\references{
\url{https://www.sfu.ca/~ssurjano/griewank.html}
	http://www-optima.amp.i.kyoto-u.ac.jp/member/student/hedar/Hedar_files/TestGO.htm.
	Molga, M., & Smutnicki, C. Test functions for optimization needs (2005). Retrieved June 2013, from http://www.zsd.ict.pwr.wroc.pl/files/docs/functions.pdf.
}
\author{
Sonja Surjanovic, Simon Fraser University
         	Derek Bingham, Simon Fraser University
}
