% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grlee12.R
\name{grlee12}
\alias{grlee12}
\title{GRAMACY & LEE (2012) FUNCTION}
\usage{
grlee12(x)
}
\description{
Dimensions: 1
	This is a simple one-dimensional test function.
	This function is usually evaluated on the x <U+2208> [0.5, 2.5].
}
\references{
\url{https://www.sfu.ca/~ssurjano/grlee12.html}
	Gramacy, R. B., & Lee, H. K. (2012). Cases for the nugget in modeling computer experiments. Statistics and Computing, 22(3), 713-722.
	Ranjan, P. (2013). Comment: EI Criteria for Noisy Computer Simulators. Technometrics, 55(1), 24-28.
}
\author{
Sonja Surjanovic, Simon Fraser University
         	Derek Bingham, Simon Fraser University
}
