% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hart3.R
\name{hart3}
\alias{hart3}
\title{HARTMANN 3-DIMENSIONAL FUNCTION}
\usage{
hart3(xx)
}
\arguments{
\item{xx}{c(x1, x2, x3)}
}
\description{
Dimensions: 3
	The 3-dimensional Hartmann function has 4 local minima.
	The function is usually evaluated on the hypercube xi <U+2208> (0, 1), for all i = 1, 2, 3.
	Reference:
	http://www-optima.amp.i.kyoto-u.ac.jp/member/student/hedar/Hedar_files/TestGO.htm.
}
\author{
Sonja Surjanovic, Simon Fraser University
	Derek Bingham, Simon Fraser University
}
