% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hart4.R
\name{hart4}
\alias{hart4}
\title{HARTMANN 4-DIMENSIONAL FUNCTION}
\usage{
hart4(xx)
}
\arguments{
\item{xx}{c(x1, x2, x3, x4)}
}
\description{
Dimensions: 4
	The 4-dimensional Hartmann function is multimodal. It is given here in the form of Picheny et al. (2012), having a mean of zero and a variance of one. The authors also add a small Gaussian error term to the output.
	The function is evaluated on the hypercube xi <U+2208> [0, 1], for all i = 1, 2, 3, 4.
}
\references{
\url{https://www.sfu.ca/~ssurjano/hart4.html}
	Dixon, L. C. W., & Szego, G. P. (1978). The global optimization problem: an introduction. Towards global optimization, 2, 1-15.
	Picheny, V., Wagner, T., & Ginsbourger, D. (2012). A benchmark of kriging-based infill criteria for noisy optimization.
}
\author{
Sonja Surjanovic, Simon Fraser University
         	Derek Bingham, Simon Fraser University
}
