% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hart6.R
\name{hart6}
\alias{hart6}
\title{HARTMANN 6-DIMENSIONAL FUNCTION}
\usage{
hart6(xx)
}
\arguments{
\item{xx}{c(x1, x2, x3, x4, x5, x6)}
}
\description{
Dimensions: 6
	The 6-dimensional Hartmann function has 6 local minima.
	The function is usually evaluated on the hypercube xi <U+2208> (0, 1), for all i = 1, ?, 6.
	Modifications and Alternate Forms:
	Picheny et al. (2012) use the following rescaled form of the 6-dimensional Hartmann function:
	This rescaled form of the function has a mean of zero and a variance of one. The authors also add a small Gaussian error term to the output.
}
\references{
\url{https://www.sfu.ca/~ssurjano/hart6.html}
	Dixon, L. C. W., & Szego, G. P. (1978). The global optimization problem: an introduction. Towards global optimization, 2, 1-15.
	http://www-optima.amp.i.kyoto-u.ac.jp/member/student/hedar/Hedar_files/TestGO.htm.
	Picheny, V., Wagner, T., & Ginsbourger, D. (2012). A benchmark of kriging-based infill criteria for noisy optimization.
}
\author{
Sonja Surjanovic, Simon Fraser University
         	Derek Bingham, Simon Fraser University
}
