% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holder.R
\name{holder}
\alias{holder}
\title{HOLDER TABLE FUNCTION}
\usage{
holder(xx)
}
\arguments{
\item{xx}{c(x1, x2)}
}
\description{
Dimensions: 2
	The Holder Table function has many local minima, with four global minima.
	The function is usually evaluated on the square xi <U+2208> [-10, 10], for all i = 1, 2.
}
\references{
\url{https://www.sfu.ca/~ssurjano/holder.html}
	Test functions for optimization. In Wikipedia. Retrieved June 2013, from https://en.wikipedia.org/wiki/Test_functions_for_optimization.
}
\author{
Sonja Surjanovic, Simon Fraser University
         	Derek Bingham, Simon Fraser University
}
