% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levy.R
\name{levy}
\alias{levy}
\title{LEVY FUNCTION}
\usage{
levy(xx)
}
\arguments{
\item{xx}{c(x1, x2, ..., xd)}
}
\description{
Dimensions: d
	The function is usually evaluated on the hypercube xi <U+2208> [-10, 10], for all i = 1, ?, d.
	Reference:
	Laguna, M., & Marti, R. Experimental Testing of Advanced Scatter Search Designs for Global Optimization of Multimodal Functions (2002). Retrieved June 2013, from http://www.uv.es/rmarti/paper/docs/global1.pdf.
}
\author{
Sonja Surjanovic, Simon Fraser University
         	Derek Bingham, Simon Fraser University
}
