% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levy13.R
\name{levy13}
\alias{levy13}
\title{LEVY FUNCTION N. 13}
\usage{
levy13(xx)
}
\arguments{
\item{xx}{c(x1, x2)}
}
\description{
Dimensions: 2
	The function is usually evaluated on the square xi <U+2208> [-10, 10], for all i = 1, 2.
	Reference:
}
\author{
Sonja Surjanovic, Simon Fraser University
         	Derek Bingham, Simon Fraser University
}
