% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{liverDisorders}
\alias{liverDisorders}
\alias{bupa}
\title{Liver Disorders}
\format{
A data frame with 345 observations and 7 features
}
\source{
Liver Disorders. (1990).
 UCI Machine Learning Repository. https://doi.org/10.24432/C54G67.
}
\usage{
data(liverDisorders)
}
\description{
The first 5 variables are all blood tests which are thought to
be sensitive to liver disorders that might arise from excessive alcohol
consumption. Each line in the dataset constitutes the record of a single
 male individual.
Important note: The 7th field (selector) has been widely misinterpreted
in the past as a dependent variable representing presence or absence of
 a liver disorder. This is incorrect [1].
   The 7th field was created by BUPA researchers as a train/test selector.
   It is not suitable as a dependent variable for classification.
   The dataset does not contain any variable representing presence or
   absence of a liver disorder. Researchers who wish to use this
   dataset as a classification benchmark should follow the method
   used in experiments by the donor (Forsyth & Rada, 1986, Machine
  learning: applications in expert systems and information retrieval)
  and others (e.g. Turney, 1995, Cost-sensitive classification:
  Empirical evaluation of a hybrid genetic decision tree induction
  algorithm), who used the 6th field (drinks), after dichotomising,
  as a dependent variable for classification. Because of widespread
  misinterpretation in the past, researchers should take care
  to state their method clearly.
}
\details{
\describe{
\item{\code{selector}}{The response variable. field created by the BUPA
researchers to split the data into train/test sets}
\item{\code{mcv}}{mean corpuscular volume}
\item{\code{alkphos}}{alkaline phosphotase}
\item{\code{sgpt}}{alanine aminotransferase}
\item{\code{sgot}}{aspartate aminotransferase}
\item{\code{gammagt}}{gamma-glutamyl transpeptidase}
\item{\code{drinks}}{number of half-pint equivalents of alcoholic beverages
 drunk per day}
}
}
