% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{maternalHealth}
\alias{maternalHealth}
\title{Maternal Health Risk}
\format{
A data frame with 1014 observations and 7 features
}
\source{
Ahmed,Marzia. (2023). Maternal Health Risk. UCI Machine Learning
Repository. https://doi.org/10.24432/C5DP5D.
}
\usage{
data(maternalHealth)
}
\description{
Data has been collected from different hospitals, community
clinics, maternal health cares from the rural areas of Bangladesh
through the IoT based risk monitoring system.
}
\details{
\describe{
\item{\code{RiskLevel}}{The response variable with 3 levels -- high,
low and mid}
\item{\code{Age}}{}
\item{\code{SystolicBP}}{Systolic Blood Pressure}
\item{\code{DiastolicBP}}{Diastolic Bloop Pressure}
\item{\code{BS}}{Blood Sugar}
\item{\code{BodyTemp}}{Body Temperature}
\item{\code{HeartRate}}{aspartate aminotransferase}
}
}
