% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matya.R
\name{matya}
\alias{matya}
\title{MATYAS FUNCTION}
\usage{
matya(xx)
}
\arguments{
\item{xx}{c(x1, x2)}
}
\description{
Dimensions: 2
	The Matyas function has no local minima except the global one.
	The function is usually evaluated on the square xi <U+2208> [-10, 10], for all i = 1, 2.
	Reference:
	http://www-optima.amp.i.kyoto-u.ac.jp/member/student/hedar/Hedar_files/TestGO.htm.
}
\author{
Sonja Surjanovic, Simon Fraser University
         	Derek Bingham, Simon Fraser University
}
