% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mccorm.R
\name{mccorm}
\alias{mccorm}
\title{MCCORMICK FUNCTION}
\usage{
mccorm(xx)
}
\arguments{
\item{xx}{c(x1, x2)}
}
\description{
Dimensions: 2
	The function is usually evaluated on the rectangle x1 <U+2208> [-1.5, 4], x2 <U+2208> [-3, 4].
	Reference:
	Adorio, E. P., & Diliman, U. P. MVF - Multivariate Test Functions Library in C for Unconstrained Global Optimization (2005). Retrieved June 2013, from http://http://www.geocities.ws/eadorio/mvf.pdf.
}
\author{
Sonja Surjanovic, Simon Fraser University
         	Derek Bingham, Simon Fraser University
}
