% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michal.R
\name{michal}
\alias{michal}
\title{MICHALEWICZ FUNCTION}
\usage{
michal(xx, m = 10)
}
\arguments{
\item{xx}{c(x1, x2)}

\item{m}{constant (optional), with default value 10}
}
\description{
Dimensions: d
	The Michalewicz function has d! local minima, and it is multimodal. The parameter m defines the steepness of they valleys and ridges; a larger m leads to a more difficult search. The recommended value of m is m = 10. The function's two-dimensional form is shown in the plot above.
	The function is usually evaluated on the hypercube xi <U+2208> [0, p], for all i = 1, ?, d.
}
\references{
\url{https://www.sfu.ca/~ssurjano/michal.html}
	http://www-optima.amp.i.kyoto-u.ac.jp/member/student/hedar/Hedar_files/TestGO.htm.
	Molga, M., & Smutnicki, C. Test functions for optimization needs (2005). Retrieved June 2013, from http://www.zsd.ict.pwr.wroc.pl/files/docs/functions.pdf.
}
\author{
Sonja Surjanovic, Simon Fraser University
         	Derek Bingham, Simon Fraser University
}
