% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{my_contour}
\alias{my_contour}
\title{contour plots}
\usage{
my_contour(
  x1,
  x2,
  Z,
  xlab = NULL,
  ylab = NULL,
  title = NULL,
  subtitle = "it1",
  background = "white",
  title.adjust = 0.5,
  nlevels = 100,
  col = "blue",
  ...
)
}
\arguments{
\item{x1, x2}{locations of grid lines at which the values in z are measured.
These must be in ascending order. By default, equally spaced values from 0 to 1 are used.
If x is a list, its components x$x and x$y are used for x and y, respectively.
If the list has component z this is used for z.}

\item{Z}{a matrix containing the values to be plotted
(NAs are allowed). Note that x can be used instead of z
for convenience.}

\item{xlab, ylab}{labels for x and y axis respectively. defaults to NULL}

\item{title, subtitle}{title/subtitle of the plot}

\item{background}{the background color}

\item{title.adjust}{adjust the title}

\item{nlevels}{number of contour levels desired *iff* levels is not supplied.}

\item{col}{the color for the contours}

\item{...}{additional arguments to plot.window, title, Axis and box,
typically graphical parameters such as cex.axis.}
}
\description{
Uses EGO algorithm to optimize any given function.
}
\author{
BLANK
}
