% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{parkinsons}
\alias{parkinsons}
\title{Maternal Health Risk}
\format{
A data frame with 195 observations and 23 features
}
\source{
Ahmed,Marzia. (2023). Maternal Health Risk. UCI Machine Learning
Repository. https://doi.org/10.24432/C5DP5D.
}
\usage{
data(parkinsons)
}
\description{
This dataset is composed of a range of biomedical voice
measurements from 31 people, 23 with Parkinson's disease (PD).
Each column in the table is a particular voice measure, and each row
corresponds one of 195 voice recording from these
individuals ("name" column). The main aim of the data is to discriminate
healthy people from those with PD, according to "status" column
 which is set to 0 for healthy and 1 for PD.
}
\details{

}
