% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm0db.R
\name{perm0db}
\alias{perm0db}
\title{PERM FUNCTION 0, d, beta}
\usage{
perm0db(xx, b = 10)
}
\arguments{
\item{xx}{c(x1, x2)}

\item{b}{constant (optional), with default value 10}
}
\description{
Dimensions: d
	The function is usually evaluated on the hypercube xi <U+2208> [-d, d], for all i = 1, ?, d.
	Reference:
	http://www-optima.amp.i.kyoto-u.ac.jp/member/student/hedar/Hedar_files/TestGO.htm.
}
\author{
Sonja Surjanovic, Simon Fraser University
         	Derek Bingham, Simon Fraser University
}
