% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permdb.R
\name{permdb}
\alias{permdb}
\title{PERM FUNCTION d, beta}
\usage{
permdb(xx, b = 0.5)
}
\arguments{
\item{xx}{c(x1, x2)
b  = constant (optional), with default value 0.5}
}
\description{
Dimensions: d
	The Perm d, ? function.
	The function is usually evaluated on the hypercube xi <U+2208> [-d, d], for all i = 1, ?, d.
	Reference:
	http://www-optima.amp.i.kyoto-u.ac.jp/member/student/hedar/Hedar_files/TestGO.htm.
}
\author{
Sonja Surjanovic, Simon Fraser University
         	Derek Bingham, Simon Fraser University
}
