% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotComparison}
\alias{plotComparison}
\title{plot the compared methods}
\usage{
plotComparison(
  res,
  n = NULL,
  maximize = FALSE,
  m = 0,
  nsteps = 5,
  errorbars = TRUE,
  which = "errors"
)
}
\description{
Uses EGO algorithm to optimize any given function.
}
\author{
BLANK
}
