% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powell.R
\name{powell}
\alias{powell}
\title{POWELL FUNCTION}
\usage{
powell(xx)
}
\arguments{
\item{xx}{c(x1, x2, ..., xd)}
}
\description{
Dimensions: d
	The function is usually evaluated on the hypercube xi <U+2208> [-4, 5], for all i = 1, ?, d.
	Reference:
	http://www-optima.amp.i.kyoto-u.ac.jp/member/student/hedar/Hedar_files/TestGO.htm.
	Laguna, M., & Marti, R. Experimental Testing of Advanced Scatter Search Designs for Global Optimization of Multimodal Functions (2002). Retrieved June 2013, from http://www.uv.es/rmarti/paper/docs/global1.pdf.
}
\author{
Sonja Surjanovic, Simon Fraser University
         	Derek Bingham, Simon Fraser University
}
