% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powersum.R
\name{powersum}
\alias{powersum}
\title{POWER SUM FUNCTION}
\usage{
powersum(xx, b)
}
\arguments{
\item{xx}{c(x1, x2)}

\item{b}{= d-dimensional vector (optional), with default value
c(8, 18, 44, 114) (when d=4)}
}
\description{
Dimensions: d
	The Power Sum function. It is shown above in its two-dimensional form. The recommended value of the b-vector, for d = 4, is: b = (8, 18, 44, 114).
	The function is usually evaluated on the hypercube xi <U+2208> [0, d], for all i = 1, ?, d.
	Reference:
	http://www-optima.amp.i.kyoto-u.ac.jp/member/student/hedar/Hedar_files/TestGO.htm.
}
\author{
Sonja Surjanovic, Simon Fraser University
         	Derek Bingham, Simon Fraser University
}
