% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_fun.R
\name{print.egoOptim}
\alias{print.egoOptim}
\title{print.egoOptim
print an egoOptim object}
\usage{
\method{print}{egoOptim}(x, ...)
}
\arguments{
\item{x}{an object of class egoOptim}
}
\description{
print.egoOptim
print an egoOptim object
}
