% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{raisin}
\alias{raisin}
\title{Raisin}
\source{
Çinar,İ̇lkay, Koklu,Murat, and Tasdemir,Sakir. (2023). Raisin.
 UCI Machine Learning Repository. https://doi.org/10.24432/C5660T.

 @format A data set with 900 observations and 8 features
}
\description{
Images of Kecimen and Besni raisin varieties grown in Turkey were obtained
 with CVS. A total of 900 raisin grains were used, including 450 pieces
 from both varieties. These images were subjected to various stages of
 pre-processing and 7 morphological features were extracted. These features
  have been classified using three different artificial intelligence
  techniques.
}
