% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rothyp.R
\name{rothyp}
\alias{rothyp}
\title{ROTATED HYPER-ELLIPSOID FUNCTION}
\usage{
rothyp(xx)
}
\arguments{
\item{xx}{c(x1, x2, ..., xd)}
}
\description{
Dimensions: d
	The Rotated Hyper-Ellipsoid function is continuous, convex and unimodal. It is an extension of the Axis Parallel Hyper-Ellipsoid function, also referred to as the Sum Squares function. The plot shows its two-dimensional form.
	The function is usually evaluated on the hypercube xi <U+2208> [-65.536, 65.536], for all i = 1, ?, d.
	Reference:
	Molga, M., & Smutnicki, C. Test functions for optimization needs (2005). Retrieved June 2013, from http://www.zsd.ict.pwr.wroc.pl/files/docs/functions.pdf.
}
\author{
Sonja Surjanovic, Simon Fraser University
         	Derek Bingham, Simon Fraser University
}
