% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schaffer4.R
\name{schaffer4}
\alias{schaffer4}
\title{SCHAFFER FUNCTION N. 4}
\usage{
schaffer4(xx)
}
\arguments{
\item{xx}{c(x1, x2)}
}
\description{
Dimensions: 2
	The fourth Schaffer function. It is shown on a smaller input domain in the second plot to show detail.
	The function is usually evaluated on the square xi <U+2208> [-100, 100], for all i = 1, 2.
	Reference:
	Test functions for optimization. In Wikipedia. Retrieved June 2013, from https://en.wikipedia.org/wiki/Test_functions_for_optimization.
}
\author{
Sonja Surjanovic, Simon Fraser University
         	Derek Bingham, Simon Fraser University
}
