% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schwef.R
\name{schwef}
\alias{schwef}
\title{SCHWEFEL FUNCTION}
\usage{
schwef(xx)
}
\arguments{
\item{xx}{c(x1, x2, ..., xd)}
}
\description{
Dimensions: d
	The Schwefel function is complex, with many local minima. The plot shows the two-dimensional form of the function.
	The function is usually evaluated on the hypercube xi <U+2208> [-500, 500], for all i = 1, ?, d.
	Reference:
	GEATbx: Examples of Objective Functions. Retrieved September 2014, from http://www.pg.gda.pl/~mkwies/dyd/geadocu/fcnfun7.html.
	http://www-optima.amp.i.kyoto-u.ac.jp/member/student/hedar/Hedar_files/TestGO.htm.
	Laguna, M., & Marti, R. Experimental Testing of Advanced Scatter Search Designs for Global Optimization of Multimodal Functions (2002). Retrieved June 2013, from http://www.uv.es/rmarti/paper/docs/global1.pdf.
}
\author{
Sonja Surjanovic, Simon Fraser University
         	Derek Bingham, Simon Fraser University
}
