% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{seeds}
\alias{seeds}
\title{seeds}
\format{
A dataset with 210 observations and 8 features
}
\source{
Charytanowicz,Magorzata, Niewczas,Jerzy, Kulczycki,Piotr,
Kowalski,Piotr, and Lukasik,Szymon. (2012). seeds. UCI Machine Learning
Repository. https://doi.org/10.24432/C5H30K.
}
\description{
Measurements of geometrical properties of kernels belonging to three
different varieties of wheat. A soft X-ray technique and GRAINS package
were used to construct all seven, real-valued attributes.
}
\details{
The examined group comprised kernels belonging to three different varieties
of wheat: Kama, Rosa and Canadian, 70 elements each, randomly selected for
the experiment. High quality visualization of the internal kernel structure
was detected using a soft X-ray technique. It is non-destructive and
considerably cheaper than other more sophisticated imaging techniques
like scanning microscopy or laser technology. The images were recorded on
13x18 cm X-ray KODAK plates. Studies were conducted using combine harvested
wheat grain originating from experimental fields, explored at the Institute
of Agrophysics of the Polish Academy of Sciences in Lublin.
\describe{
To construct the data, seven geometric parameters of wheat kernels were
measured: \code{area A}; \code{perimeter P};
 \code{compactness C = 4*pi*A/P^2}; \code{kernel_length} - Length of kernel;
\code{kernel_width} - width of kernel; \code{asymmetry coefficient};
\code{kernel_groove_length} - length of kernel groove.
All of these parameters were real-valued continuous.
}
}
