% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shekel.R
\name{shekel}
\alias{shekel}
\title{SHEKEL FUNCTION}
\usage{
shekel(xx, m = 10)
}
\arguments{
\item{xx}{c(x1, x2, x3, x4)}
}
\description{
Dimensions: 4
	The Shekel function has m local minima. Above are the recommended values of m, the ?-vector and the C-matrix; ? is an m-dimensional vector, and C is a 4-by-m-dimensional matrix.
	The function is usually evaluated on the hypercube xi <U+2208> [0, 10], for all i = 1, 2, 3, 4.
	Reference:
	http://www-optima.amp.i.kyoto-u.ac.jp/member/student/hedar/Hedar_files/TestGO.htm.
	Molga, M., & Smutnicki, C. Test functions for optimization needs (2005). Retrieved June 2013, from http://www.zsd.ict.pwr.wroc.pl/files/docs/functions.pdf.
}
\author{
Sonja Surjanovic, Simon Fraser University
         	Derek Bingham, Simon Fraser University
}
