% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shubert.R
\name{shubert}
\alias{shubert}
\title{SHUBERT FUNCTION}
\usage{
shubert(xx)
}
\arguments{
\item{xx}{c(x1, x2)}
}
\description{
Dimensions: 2
	The Shubert function has several local minima and many global minima. The second plot shows the the function on a smaller input domain, to allow for easier viewing.
	The function is usually evaluated on the square xi <U+2208> [-10, 10], for all i = 1, 2, although this may be restricted to the square xi <U+2208> [-5.12, 5.12], for all i = 1, 2.
	Reference:
	http://www-optima.amp.i.kyoto-u.ac.jp/member/student/hedar/Hedar_files/TestGO.htm.
}
\author{
Sonja Surjanovic, Simon Fraser University
         	Derek Bingham, Simon Fraser University
}
