% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stybtang.R
\name{stybtang}
\alias{stybtang}
\title{STYBLINSKI-TANG FUNCTION}
\usage{
stybtang(xx)
}
\arguments{
\item{xx}{c(x1, x2, ..., xd)}
}
\description{
Dimensions: d
	The Styblinski-Tang function is shown here in its two-dimensional form.
	The function is usually evaluated on the hypercube xi <U+2208> [-5, 5], for all i = 1, ?, d.
	Reference:
}
\author{
Sonja Surjanovic, Simon Fraser University
         	Derek Bingham, Simon Fraser University
}
