% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumpow.R
\name{sumpow}
\alias{sumpow}
\title{SUM OF DIFFERENT POWERS FUNCTION}
\usage{
sumpow(xx)
}
\arguments{
\item{xx}{c(x1, x2, ..., xd)}
}
\description{
Dimensions: d
	The Sum of Different Powers function is unimodal. It is shown here in its two-dimensional form.
	The function is usually evaluated on the hypercube xi <U+2208> [-1, 1], for all i = 1, ?, d.
	Reference:
	Molga, M., & Smutnicki, C. Test functions for optimization needs (2005). Retrieved June 2013, from http://www.zsd.ict.pwr.wroc.pl/files/docs/functions.pdf.
}
\author{
Sonja Surjanovic, Simon Fraser University
         	Derek Bingham, Simon Fraser University
}
