% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trid.R
\name{trid}
\alias{trid}
\title{TRID FUNCTION}
\usage{
trid(xx)
}
\arguments{
\item{xx}{c(x1, x2, ..., xd)}
}
\description{
Dimensions: d
	The Trid function has no local minimum except the global one. It is shown here in its two-dimensional form.
	The function is usually evaluated on the hypercube xi <U+2208> [-d2, d2], for all i = 1, ?, d.
}
\references{
\url{https://www.sfu.ca/~ssurjano/trid.html}
	Adorio, E. P., & Diliman, U. P. MVF - Multivariate Test Functions Library in C for Unconstrained Global Optimization (2005). Retrieved August 2017, from http://www.geocities.ws/eadorio/mvf.pdf.
	http://www-optima.amp.i.kyoto-u.ac.jp/member/student/hedar/Hedar_files/TestGO.htm.
}
\author{
Sonja Surjanovic, Simon Fraser University
         	Derek Bingham, Simon Fraser University
}
