% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{vertebral}
\alias{vertebral}
\alias{vertebral2}
\alias{vertebral3}
\title{Vertebral Column}
\format{
A dataset with 310 observations and 7 features
}
\source{
Barreto,Guilherme and Neto,Ajalmar. (2011). Vertebral Column.
 UCI Machine Learning Repository. https://doi.org/10.24432/C5K89B.

Barreto,Guilherme and Neto,Ajalmar. (2011). Vertebral Column.
UCI Machine Learning Repository. https://doi.org/10.24432/C5K89B.
}
\description{
Data set containing values for six biomechanical features used to classify
orthopaedic patients into 3 classes (normal, disk hernia or
spondilolysthesis) or 2 classes (normal or abnormal).
}
\details{
Biomedical data set built by Dr. Henrique da Mota during a medical residence
 period in the Group of Applied Research in Orthopaedics (GARO) of the
 Centre Médico-Chirurgical de Réadaptation des Massues, Lyon, France. The
  data have been organized in two different but related classification tasks.
   The first task consists in classifying patients as belonging to one out
   of three categories: Normal (100 patients), Disk Hernia (60 patients) or
   Spondylolisthesis  (150 patients). For the second task, the categories
   Disk Hernia and Spondylolisthesis were merged into a single category
   labelled as 'abnormal'. Thus, the second task consists in classifying
   patients as belonging to one out of two categories: Normal (100 patients)
    or Abnormal (210 patients). We provide files also for use within the
    WEKA environment.
\describe{
Each patient is represented in the data set by six biomechanical attributes
derived from the shape and orientation of the pelvis and lumbar spine
(in this order): pelvic incidence, pelvic tilt, lumbar lordosis angle,
sacral slope, pelvic radius and grade of spondylolisthesis.
The following convention is used for the class labels: DH (Disk Hernia),
Spondylolisthesis (SL), Normal (NO) and Abnormal (AB).
}
}
