% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{wine}
\alias{wine}
\title{wine}
\format{
A dataset of 174 observations with 14 features
}
\description{
Using chemical analysis determine the origin of wines
}
\details{
These data are the results of a chemical analysis of wines grown in the same
 region in Italy but derived from three different cultivars. The analysis
 determined the quantities of 13 constituents found in each of the three
 types of wines. I think that the initial data set had around 30 variables,
 but for some reason I only have the 13 dimensional version.
 I had a list of what the 30 or so variables were, but a.)
 I lost it, and b.), I would not know which 13 variables are included in
 the set.
\describe{
The attributes are (dontated by \code{Riccardo Leardi,
riclea@anchem.unige.it}):

\code{Alcohol; Malic acid; Ash; Alcalinity of ash; Magnesium; Total phenols;
 Flavanoids; Nonflavanoid phenols; Proanthocyanins; Color intensity; Hue;
 OD280/OD315 of diluted wines; Proline}
 }

 @source Aeberhard,Stefan and Forina,M.. (1991). Wine.
 UCI Machine Learning Repository. https://doi.org/10.24432/C5PC7J.
}
