% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zakharov.R
\name{zakharov}
\alias{zakharov}
\title{ZAKHAROV FUNCTION}
\usage{
zakharov(xx)
}
\arguments{
\item{xx}{c(x1, x2, ..., xd)}
}
\description{
Dimensions: d
	The Zakharov function has no local minima except the global one. It is shown here in its two-dimensional form.
	The function is usually evaluated on the hypercube xi <U+2208> [-5, 10], for all i = 1, ?, d.
	Reference:
	http://www-optima.amp.i.kyoto-u.ac.jp/member/student/hedar/Hedar_files/TestGO.htm.
}
\author{
Sonja Surjanovic, Simon Fraser University
         	Derek Bingham, Simon Fraser University
}
