% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_budget}
\alias{add_budget}
\title{add_budget}
\usage{
add_budget(object, its, new_budget)
}
\arguments{
\item{object}{@param object an egoOptim object or a list of egoOptim objects}

\item{its}{Integer. Number of iterations to add}

\item{new_budget}{Additional number of function evaluations}
}
\description{
A function to add budget on the evaluated object
}
