% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{method_compare}
\alias{method_compare}
\title{Function to compare different methods}
\usage{
method_compare(
  fun,
  lower,
  upper,
  ...,
  p = NULL,
  maximize = FALSE,
  reps = 20L,
  file = NULL,
  control = list(),
  overwrite = FALSE
)
}
\arguments{
\item{fun}{function to be optimized}

\item{lower}{vector of lower bounds for the variables to be optimized over,}

\item{upper}{vector of upper bounds for the variables to be optimized over,}

\item{...}{additional arguments to be passed to fun}

\item{p}{Integer. The dimension of the function in case lower is not given.}

\item{maximize}{logical. Should we maximize or minimize (the default)?}

\item{reps}{Integer, number of replicates/trials.}

\item{file}{a character string specifying where the results will be stored
in the hard drive.}

\item{control}{a list of control parameters passed to the optimize_fun.
See ‘optimized_fun’ for more information.}
}
\description{
Function to compare different methods
}
\details{
\describe{
 \item{\code{nsteps}}{A positive integer representing the
 desired number of additional points per iteration}
 \item{\code{trueglobal}}{double. The true function optimum.}
 \item{\code{expansion_rate}}{double [0,1] for the ROI expansion in case of a failed iteration. 0 implies reverting back to
the initial function domain.}
 \item{\code{tol}}{double. the desired accuracy level.ie convergence tolerance.
 If reached before maxit, the code stops and the result returned}
 \item{\code{budget}}{The total number of function evaluations to be carried out}
 \item{\code{maxit}}{Integer. maximum number of iterations to be carried out. Defaults to 20}
 \item{\code{counter}}{Integer. Number of times to switch to global search once local search is unsuccessful}
 \item{\code{do_maxit}}{logical. Whether to iterate until maxit}
 \item{\code{plot}}{logical. Should the Region be ploted?}
 \item{\code{dimplot}}{an integer vector of length 2. The dimensions of the function to be plotted}
 \item{\code{basicEGO}}{logical. Whether to maintain the domain as given.
 TRUE denotes that the domain will not change throughout the optimization. This is the basic EGO algorithm}
 \item{\code{method}}{choise between \code{fastEGO} and \code{TREGO}}
 \item{\code{n}}{number of observations for the initial design. defaults to 5*length(lower)
 Used if X is not provided. Defaults to \code{5*length(lower)}}
 \item{\code{cost}}{cost function to evaluate the loss}
 \item{\code{trace}}{logical. Tracing the information on the progress of the optimization}
 \item{\code{seed}}{Set a seed for reproducability}
}
}
\author{
BLANK
}
