% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{method_compare}
\alias{method_compare}
\title{Function to compare different methods}
\usage{
method_compare(
  fun,
  lower,
  upper,
  ...,
  budget = 70,
  p = NULL,
  maximize = FALSE,
  reps = 20L,
  expansion_rate = 0,
  file = NULL,
  control = list()
)
}
\arguments{
\item{fun}{function to be optimized}

\item{lower}{vector of lower bounds for the variables to be optimized over,}

\item{upper}{vector of upper bounds for the variables to be optimized over,}

\item{...}{additional arguments to be passed to fun}

\item{budget}{The total number of function evaluations to be carried out}

\item{p}{Integer. The dimension of the function in case lower is not given.}

\item{maximize}{logical. Should we maximize or minimize (the default)?}

\item{reps}{Integer, number of replicates/trials.}

\item{expansion_rate}{double [0,1]. For the ROI expansion in case of a
failed iteration. 0 implies reverting back to
the initial function domain.}

\item{file}{a character string specifying where the results will be stored
in the hard drive.}

\item{control}{a list of control parameters passed to the optimize_fun.
See ‘optimized_fun’ for more information.}
}
\description{
Function to compare different methods
}
\author{
BLANK
}
