% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_fun.R
\name{optimize_fun}
\alias{optimize_fun}
\title{Optimize fun}
\usage{
optimize_fun(FUN, lower, upper, n = 5, nsteps = 2, top = 0.15, ntimes = 5)
}
\description{
Uses EGO algotiyhm to optimize any given function.
}
\author{
BLANK
}
